Refactor source analysis into record
